backend/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==21.2.0
whitenoise==6.7.0   # (선택) 정적파일 필요시

